moved things around in order to remove legacy code and make the components folder more modular
