--- desktop/src/components/tables/Generic/index.js
@@ -1,2 +1,0 @@
-import Generic from './genericTable';
-export default Generic;
