+++ desktop/src/components/tables/Table/index.js
@@ -1,0 +1,2 @@
+import Table from './table';
+export default  Table;
