+++ desktop/src/components/tables/Toolbar/index.js
@@ -1,0 +1,2 @@
+import Toolbar from './toolbar';
+export default Toolbar;
