--- desktop/src/components/tables/VirtualizedSortSelect/index.js
@@ -1,2 +1,0 @@
-import VirtualizedSortSelect from './virtualizedSortSelect';
-export default  VirtualizedSortSelect;
