--- desktop/src/components/tables/toolbars/AnalyzeToolbar/index.js
@@ -1,2 +1,0 @@
-import AnalyzeToolbar from './analyzeToolbar';
-export default AnalyzeToolbar;
