+++ desktop/src/containers/Task/taskIndex.container.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import VirtualizedSortSelect from 'components/tables/VirtualizedSortSelect';
+import VirtualizedSortSelect from 'components/tables/Table';
 import Progress from 'components/helpers/Progress';
 
 import { analyzeActions } from 'store/actions';
